refactor: ChatMain
